fix(nx-plugin): Handle testing in watch mode (#208)
## Summary
Currently, when running `test:watch` or `test:ui`, prerequisite tasks
(ie, building dependencies) are not run. This adjusts both of these
commands to be set up with nx spawn and adjusts the nx preset to spin up
`build:watch` tasks for any dependencies.

## Implementation Notes
- The actual `test:watch` and `test:ui` commands do not have the build
dependencies set - they're set on the command called by `nx-spawn` so
that `nx-spawn` starts those, not `nx` itself
- `build:watch` for dependencies is the only prerequisite set of tasks.
- Caching is neither enabled nor configured for these test commands as
it doesn't make sense to cache a watch task

## Testing
Spun up two packages, one which is the dependency of the other. Making a
change in the dependency while `test:watch` was run on the main package
triggered it to be rebuilt. Note that the tests do not appear to be
automatically rerun on dependency change - I tried adjusting the vitest
configuration, but no combination of adjustments to the watchExclude or
deps parameters changed this behavior. For now, the workaround is to
just type `a` in the console to trigger a full rerun. Cypress may not
have this limitation.