fix(nx-plugin): Fix package generation with empty copyright holder (#205)
When generating a repository and specifying not to include a license or
using a custom license, the copyright holder field is not needed and
will not be set. However, when generating a package, it was defaulted to
null and is unconditionally checked to be a string. The new behavior
properly allows it to be omitted in these situations.