fix(nx-plugin): Avoid CI failure when no release PR exists (#206)
Currently, if a commit is pushed to main when the release PR branch does
not exist and is not caused to be created (eg, when committing to a repo
before any packages are configured), the CI still attempts to run
lockfile updates on the release PR branch, as we always do it whenever a
release is not created. This also prevents it from running when there is
no active release PR (when a release PR exists, the release branch must
also exist).

This was tested on a separate independent repo. Ensured the job ran when
new commits are pushed to the release PR, but not when the release PR is
not present or when merging the release PR